Move out account creation from config test, and create Perps trading test file
